implemented suggestions
